Move demo render helpers into actions
