+++ .agents/skills/make-demo/SKILL.md
@@ -32,7 +32,7 @@ A good demo should:
 - Demos should push Remix to its limits in a focused way. Prefer realistic edge cases, composition, streaming, middleware, routing, navigation, forms, or request-handling scenarios over toy examples.
 - When demos use `remix/ui`, prefer idiomatic Remix component patterns. Use normal JSX composition and built-in styling/mixin props such as `css={...}` or `mix={css(...)}` and `mix={[...]}` instead of dropping down to manual DOM mutation or ad hoc class management.
 - When a demo uses `remix/ui` JSX, configure that demo's `tsconfig.json` with `jsx: "react-jsx"`, `jsxImportSource: "remix/ui"`, and `preserveSymlinks: true`. Do not add `paths` entries that point back into `packages/remix/src`. The goal is for TypeScript to resolve `remix` through the demo's own `node_modules` view, not through repo-relative source paths.
-- For HTML responses rendered with `remix/ui`, prefer a tiny local `render()` helper that calls `renderToStream(...)` and wraps it with `createHtmlResponse(...)` from `remix/response/html` instead of manually building HTML `Response` headers or wrapping the stream yourself.
+- For HTML responses rendered with `remix/ui`, prefer a tiny local `app/actions/render.tsx` helper that calls `renderToStream(...)` and wraps it with `createHtmlResponse(...)` from `remix/response/html` instead of manually building HTML `Response` headers or wrapping the stream yourself.
 - Prefer direct use of Remix and package APIs in demo code. Do not add custom wrappers around simple calls like `session.get()`, `session.set()`, `session.flash()`, `session.unset()`, `redirect()`, or `context.get(...)` unless the wrapper adds real domain logic, reusable policy, or a genuinely clearer abstraction.
 - Demo code must have good hygiene. Use clear names, small focused modules, explicit control flow, and accessible markup. Avoid hacks, dead code, unexplained shortcuts, or patterns that would be poor examples for users to copy.
 - Make the demo teach good patterns. Assume readers and future agents will study it as an example of how Remix code should be written in this repository.
