+++ demos/bookstore/README.md
@@ -31,7 +31,7 @@ Then visit http://localhost:44100
 - [`app/actions/books/controller.tsx`](app/actions/books/controller.tsx), [`app/actions/cart/controller.tsx`](app/actions/cart/controller.tsx), and [`app/actions/checkout/controller.tsx`](app/actions/checkout/controller.tsx) show the same pattern for browsing, cart mutations, and checkout flows.
 - [`app/actions/auth/controller.tsx`](app/actions/auth/controller.tsx) owns only the direct auth actions, while nested auth route maps such as login, registration, forgot-password, and reset-password are mapped explicitly in [`app/router.ts`](app/router.ts).
 - [`app/actions/admin/books/controller.tsx`](app/actions/admin/books/controller.tsx) demonstrates admin CRUD with multipart uploads while keeping forms and page rendering in feature-local `form.tsx`, `index-page.tsx`, and `show-page.tsx` files.
-- [`app/utils/render.tsx`](app/utils/render.tsx) provides the HTML renderer and preserves server-side frame resolution for fragment-based rendering.
+- [`app/actions/render.tsx`](app/actions/render.tsx) provides the controller-owned HTML renderer and preserves server-side frame resolution for fragment-based rendering.
 - [`app/ui/layout.tsx`](app/ui/layout.tsx) and [`app/ui/document.tsx`](app/ui/document.tsx) define the shared document shell, while reusable UI primitives live under `app/ui/` instead of a generic components folder.
 - [`app/utils/assets.ts`](app/utils/assets.ts) and [`app/middleware/asset-entry.ts`](app/middleware/asset-entry.ts) compile both the client entry and the demo stylesheet through `createAssetServer()`, so the demo exercises Remix's JS and CSS asset pipeline together.
 - [`app/middleware/session.ts`](app/middleware/session.ts) configures the signed cookie and filesystem-backed session storage, and [`app/middleware/uploads.ts`](app/middleware/uploads.ts) owns the upload handler used by the router.
