+++ demos/bookstore/app/actions/account/controller.tsx
@@ -3,7 +3,7 @@ import type { Controller } from 'remix/fetch-router'
 import { requireAuth } from '../../middleware/auth.ts'
 import type { routes } from '../../routes.ts'
 import { getCurrentUser } from '../../utils/context.ts'
-import { render } from '../../utils/render.tsx'
+import { render } from '../render.tsx'
 import { AccountPage } from './page.tsx'
 
 export default {
