+++ demos/bookstore/app/actions/admin/books/controller.tsx
@@ -10,7 +10,7 @@ import { requireAdmin } from '../../../middleware/admin.ts'
 import { requireAuth } from '../../../middleware/auth.ts'
 import { routes } from '../../../routes.ts'
 import { parseId } from '../../../utils/ids.ts'
-import { render } from '../../../utils/render.tsx'
+import { render } from '../../render.tsx'
 import { AdminBookFormPage } from './form.tsx'
 import { AdminBooksIndexPage } from './index-page.tsx'
 import { AdminBookNotFoundPage, AdminBookShowPage } from './show-page.tsx'
