+++ demos/bookstore/app/actions/admin/users/controller.tsx
@@ -10,7 +10,7 @@ import { requireAuth } from '../../../middleware/auth.ts'
 import { routes } from '../../../routes.ts'
 import { getCurrentUser } from '../../../utils/context.ts'
 import { parseId } from '../../../utils/ids.ts'
-import { render } from '../../../utils/render.tsx'
+import { render } from '../../render.tsx'
 import { AdminUserFormPage } from './form.tsx'
 import { AdminUsersIndexPage } from './index-page.tsx'
 import { AdminUserNotFoundPage, AdminUserShowPage } from './show-page.tsx'
