+++ demos/bookstore/app/actions/auth/forgot-password/controller.tsx
@@ -4,7 +4,7 @@ import { Database } from 'remix/data-table'
 
 import { passwordResetTokens, users } from '../../../data/schema.ts'
 import type { routes } from '../../../routes.ts'
-import { render } from '../../../utils/render.tsx'
+import { render } from '../../render.tsx'
 import { forgotPasswordSchema, normalizeEmail } from '../schemas.ts'
 import { ForgotPasswordPage, ForgotPasswordSuccessPage } from './page.tsx'
