+++ demos/bookstore/app/actions/auth/reset-password/controller.tsx
@@ -7,7 +7,7 @@ import { passwordResetTokens, users } from '../../../data/schema.ts'
 import { Session } from '../../../middleware/session.ts'
 import { routes } from '../../../routes.ts'
 import { hashPassword } from '../../../utils/password-hash.ts'
-import { render } from '../../../utils/render.tsx'
+import { render } from '../../render.tsx'
 import { resetPasswordSchema } from '../schemas.ts'
 import { ResetPasswordPage, ResetPasswordSuccessPage } from './page.tsx'
