+++ demos/bookstore/app/actions/books/controller.tsx
@@ -4,7 +4,7 @@ import { Database, ilike } from 'remix/data-table'
 import { books } from '../../data/schema.ts'
 import type { routes } from '../../routes.ts'
 import { getCurrentCart } from '../../utils/context.ts'
-import { render } from '../../utils/render.tsx'
+import { render } from '../render.tsx'
 import { IndexPage } from './index-page.tsx'
 import { GenreNotFoundPage, GenrePage } from './genre-page.tsx'
 import { BookNotFoundPage, ShowPage } from './show-page.tsx'
