+++ demos/bookstore/app/actions/contact/controller.tsx
@@ -1,7 +1,7 @@
 import type { Controller } from 'remix/fetch-router'
 
 import type { routes } from '../../routes.ts'
-import { render } from '../../utils/render.tsx'
+import { render } from '../render.tsx'
 import { ContactPage, ContactSuccessPage } from './page.tsx'
 
 export default {
