+++ demos/bookstore/app/actions/controller.tsx
@@ -6,7 +6,7 @@ import { books } from '../data/schema.ts'
 import type { routes } from '../routes.ts'
 import { assetServer } from '../utils/assets.ts'
 import { getCurrentCart } from '../utils/context.ts'
-import { render } from '../utils/render.tsx'
+import { render } from './render.tsx'
 import { uploadsStorage } from '../utils/uploads.ts'
 import { AboutPage } from './about.tsx'
 import { HomePage } from './home.tsx'
