+++ demos/frame-navigation/app/actions/main/controller.tsx
@@ -4,7 +4,7 @@ import type { RemixNode } from 'remix/ui'
 import { requireAuth } from '../../middleware/auth.ts'
 import type { routes } from '../../routes.ts'
 import { Layout, type MainNavItem } from '../../ui/layout.tsx'
-import { render } from '../../utils/render.tsx'
+import { render } from '../render.tsx'
 import { MainAccountPage } from './account-page.tsx'
 import { MainCalendarPage } from './calendar-page.tsx'
 import { MainCoursesPage } from './courses-page.tsx'
