+++ demos/social-auth/app/actions/controller.tsx
@@ -13,7 +13,7 @@ import {
   readExternalProviderLinks,
   type ExternalProviderRegistry,
 } from '../utils/external-auth.ts'
-import { render } from '../utils/render.tsx'
+import { render } from './render.tsx'
 
 export function createRootController(
   registry: ExternalProviderRegistry = externalProviderRegistry,
