+++ demos/sse/app/actions/render.ts
