+++ demos/unpkg/README.md
@@ -17,7 +17,7 @@ Semver ranges are supported (via the [semver](https://www.npmjs.com/package/semv
 ## Code Highlights
 
 - [`app/router.ts`](app/router.ts) maps route maps to controllers, matching the layout conventions used by the other Remix demos.
-- [`app/actions/controller.ts`](app/actions/controller.ts) owns the top-level home and package-browser actions.
+- [`app/actions/controller.ts`](app/actions/controller.ts) owns the top-level home and package-browser actions, with package-browser response rendering colocated in [`app/actions/package-browser.ts`](app/actions/package-browser.ts) and lower-level formatting/MIME helpers in [`app/utils`](app/utils).
 - [`app/utils/npm.ts`](app/utils/npm.ts) fetches package tarballs from npm, decompresses them with `node:zlib`, and parses them using `@remix-run/tar-parser`. The `parsePackagePath()` function handles tricky URLs like `/@remix-run/cookie@1.0.0/src/index.ts`.
 - [`app/utils/tarball-cache.ts`](app/utils/tarball-cache.ts) stores decompressed tarballs under the demo’s root `tmp/` directory using `remix/file-storage/fs`, which keeps runtime scratch data inside the demo itself.
-- [`app/utils/render.ts`](app/utils/render.ts) and [`app/ui/document.ts`](app/ui/document.ts) keep the shared document shell separate from feature controllers.
+- [`app/actions/render.ts`](app/actions/render.ts) owns HTML response creation, while [`app/ui/document.ts`](app/ui/document.ts) keeps the shared document shell separate from feature controllers.
