--- demos/unpkg/app/ui/package-browser/error.ts
@@ -1,19 +1,0 @@
-import { html } from 'remix/html-template'
-
-import { render } from '../../utils/render.ts'
-
-export function renderError(title: string, message: string): Response {
-  return render(
-    title,
-    html`
-      <h1>${title}</h1>
-      <div class="error">
-        <p>${message}</p>
-      </div>
-      <p style="margin-top: 1rem;">
-        <a href="/">Back to home</a>
-      </p>
-    `,
-    { status: 404 },
-  )
-}
