+++ template/app/actions/render.tsx
@@ -3,7 +3,7 @@ import * as path from 'node:path'
 import type { RemixNode } from 'remix/ui'
 import { renderToStream } from 'remix/ui/server'
 
-import { assetServer } from '../actions/controller.tsx'
+import { assetServer } from './controller.tsx'
 import { router } from '../router.ts'
 
 export function render(node: RemixNode, request: Request, init?: ResponseInit) {
